// back-end/src/common/enums/property.enum.ts
export enum PropertyType {
  APARTMENT = 'apartment',
  VILLA = 'villa',
  PLOT = 'plot',
  COMMERCIAL = 'commercial',
  PENTHOUSE = 'penthouse',
}

export enum PropertyStatus {
  PENDING = 'pending',
  APPROVED = 'approved',
  REJECTED = 'rejected',
  AVAILABLE = 'available',
  SOLD = 'sold',
  RENTED = 'rented',
  UNDER_OFFER = 'under_offer',
}

export enum ListingType {
  SALE = 'sale',
  RENT = 'rent',
}